wip: refactored create queue sheet
