+++ apps/studio/components/interfaces/Integrations/Queues/CreateQueueSheet/index.ts
@@ -1,0 +1,3 @@
+export { CreateQueueSheet } from './CreateQueueSheet'
+export type { CreateQueueSheetProps } from './CreateQueueSheet'
+export type { CreateQueueForm, QueueType } from './CreateQueueSheet.schema'
